+++ src/utils/round.ts
@@ -1,9 +1,13 @@
-import { ZERO_BI } from '../constants'
-
-export function roundToNearest(value: bigint, nearest: bigint): bigint {
-	if (nearest === ZERO_BI) {
-		return value
+// Fixes JS float imprecision error
+export function truncateFloat(input: number, increment: number): number {
+	const orderOfIncrement = Math.floor(Math.log10(increment))
+	if (orderOfIncrement < 0) {
+		return Number(input.toFixed(-orderOfIncrement))
+	} else {
+		return Number(input.toFixed(0))
 	}
+}
 
-	return (value / nearest) * nearest
+export function roundUpTo(initial: number, rounding: number): number {
+	return Math.ceil(initial / rounding) * rounding
 }
